requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-cors-headers==2.1.0
django-countries==4.6.2
django-filter==1.0.4
django-multiselectfield==0.1.8
djangorestframework==3.6.4
idna==2.6
Markdown==2.6.9
mysqlclient==1.3.12
python-decouple==3.1
pytz==2017.2
requests==2.18.4
spotipy==2.4.4
urllib3==1.22
